fix compilation errors and refine the implementations.
