+++ include/traits/base.hpp
@@ -51,65 +51,4 @@ struct BaseTileShape {
     static constexpr int kNumel = 256 /* kRows * kCols */;
 };
 
-/**
- * @brief The base tile shape for Swizzle<3, 3, 3>.
- */
-template <typename Element, int kBytes>
-    requires BaseType<Element>
-struct SwizzleBaseTileShape;
-
-template <typename Element>
-    requires HalfType<Element>
-struct SwizzleBaseTileShape<Element, 128> {
-    using DType = Element;
-
-    static constexpr int kRows = 8;
-    static constexpr int kCols = 64;
-    static constexpr int kNumel = kRows * kCols;
-
-    static constexpr int B = 3;
-    static constexpr int M = 3;
-    static constexpr int S = 3;
-};
-
-template <>
-struct SwizzleBaseTileShape<float, 128> {
-    using DType = float;
-
-    static constexpr int kRows = 8;
-    static constexpr int kCols = 32;
-    static constexpr int kNumel = kRows * kCols;
-
-    static constexpr int B = 3;
-    static constexpr int M = 2;
-    static constexpr int S = 3;
-};
-
-template <typename Element>
-    requires HalfType<Element>
-struct SwizzleBaseTileShape<Element, 64> {
-    using DType = Element;
-
-    static constexpr int kRows = 4;
-    static constexpr int kCols = 32;
-    static constexpr int kNumel = kRows * kCols;
-
-    static constexpr int B = 2;
-    static constexpr int M = 3;
-    static constexpr int S = 2;
-};
-
-template <>
-struct SwizzleBaseTileShape<float, 64> {
-    using DType = float;
-
-    static constexpr int kRows = 4;
-    static constexpr int kCols = 16;
-    static constexpr int kNumel = kRows * kCols;
-
-    static constexpr int B = 2;
-    static constexpr int M = 2;
-    static constexpr int S = 2;
-};
-
 }  // namespace tilefusion::traits
